Fix bugs and change name
