+++ .gitignore
@@ -1,0 +1,81 @@
+*.7z
+*.a
+*.d
+*.exe
+*.o
+*.tmp
+*.zip
+
+# Visual Studio files
+*.code-workspace
+*.ilk
+*.opensdf
+*.pdb
+*.lib
+*.sdf
+*.suo
+*.user
+*.vcxproj.filters
+.vs/
+.vscode/
+bin/
+Debug*/
+Release*/
+obj/
+
+# Temporary and backup files
+*.bak
+*.cbTemp
+*.sw?
+*~
+\#*
+gmon.out
+
+# System Files
+.DS_Store
+._.DS_Store
+Thumbs.db
+cscope.*
+tags
+
+# Netbeans IDE folders
+nbproject/
+
+# Python runtime files
+*.pyc
+
+# vim files
+*.vim
+*.session
+.ycm_extra_conf.py
+.vimrc
+
+# taskwarrior files
+*.task*
+
+
+# Transifex Client
+.tx/
+
+# Snapcraft build directories
+prime
+snap
+stage
+
+# Snapcraft build file
+*.snap
+
+# Eclipse IDE
+.cproject
+.settings/
+
+# Apple Xcode IDE
+Xcode/
+
+# gdb
+.gdb_history
+
+# Flatpak build artifacts
+/.flatpak-builder/
+/repo/
+/build-dir/
