--- ConsoleApp1/bin/x64/Debug/DECryptSuite.exe.config
@@ -1,21 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-    <configSections>
-        <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-            <section name="ConsoleApp1.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
-        </sectionGroup>
-    </configSections>
-    <startup> 
-        <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5.2" />
-    </startup>
-    <userSettings>
-        <ConsoleApp1.Properties.Settings>
-            <setting name="Rmlyc3RSdW4E" serializeAs="String">
-                <value>dHJ1ZQ==</value>
-            </setting>
-            <setting name="U2lnbgqq" serializeAs="String">
-                <value />
-            </setting>
-        </ConsoleApp1.Properties.Settings>
-    </userSettings>
-</configuration>
