+++ DECryptSuite.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26228.4
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DECryptSuite", "ConsoleApp1\DECryptSuite.csproj", "{D31C425F-FD39-4A7D-BF8A-1AACAEF70572}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DECryptSuite", "DECryptSuite\DECryptSuite.csproj", "{D31C425F-FD39-4A7D-BF8A-1AACAEF70572}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
