+++ DECryptSuite/Properties/Settings.settings
@@ -1,0 +1,12 @@
+﻿<?xml version='1.0' encoding='utf-8'?>
+<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="ConsoleApp1.Properties" GeneratedClassName="Settings">
+  <Profiles />
+  <Settings>
+    <Setting Name="Rmlyc3RSdW4E" Type="System.String" Scope="User">
+      <Value Profile="(Default)">dHJ1ZQ==</Value>
+    </Setting>
+    <Setting Name="U2lnbgqq" Type="System.String" Scope="User">
+      <Value Profile="(Default)" />
+    </Setting>
+  </Settings>
+</SettingsFile>
